renamed the Schema package to Model
